Flatten stuff out a bit
